This pull does some changes in the new sneak code:
- Variables `m_old_node_below*` and `m_need_to_get_new_sneak_node` are deprecated now
- ~Remove variable `got_teleported`~ already merged
- Remove `air` node check for expired sneak nodes
- Various code style corrections (ignore whitespaces to review)
- Compress `detectSneakLadder` and `detectSneakLedge` into four entire code lines
- Makes sneak ladders feel like climbing
- Disallow and reallow double jumps (two nodes at once) - please decide what you want
- Comparison video: https://www.youtube.com/watch?v=coTcDzEfku4 (already outdated)

Features:
- [x] ~No damage evading on sneak ladders~ already merged
- [x] ~No jitter when moving against sneak limit (#5449)~ already merged
- [ ] No weird sinking
- [x] slower 1 row sneak ladder (using #5527)
- [x] ~slower 2 row sneak ladders~ almost the same speed as in the old sneak code
- [x] No weird behaviour on [stairs](https://i.imgur.com/EtkNVPZ.png)
- [x] 8 times fewer direction checks for sneak ladders + ledges **\*1**
- [x] Jumping on node edges when sneaking (required to make sneak ladders work properly)
- [x] Walking around ledge edges, like [here](https://camo.githubusercontent.com/c132b89c7c9b7480ef84c6506a6238b01e0e586f/687474703a2f2f6936332e74696e797069632e636f6d2f786466746f6c2e6a7067) is possible (see also the video above)

**\*1**) With this pull it will only check the nodes above the sneak node, which incredibly reduces the total checks for sneak ladders & ledges down to _2_, previously up to _16_ for both **!** 

What has not changed: It is still possible to deactivate the sneak glitch or sneak itself entirely using `set_physics_override`.