@Fixer-007 2017-04-07
What I see:
+ sneak ladders are indeed slower, with speeds close to using wooden ladders
+ double jump is possible
+ no jitter at the limit of the node
+ sneaking on door is weird
+ tried some damage evasion - seen nothing new apart from https://github.com/minetest/minetest/issues/329
+ teleport to center on stairs/doors still present
+ one way sneaking + instamove on stairs still present