@SmallJoker 2017-04-07
> sneaking on door is weird

```Lua
	-- collision_box needed otherise selection box would be full node size
	collision_box = {
		type = "fixed",
		fixed = {-15/32, 13/32, -15/32, -13/32, 1/2, -13/32},
	},
```
[Doors Saucecode](https://github.com/minetest/minetest_game/blob/master/mods/doors/init.lua#L98-L102)
The new sneak code ensures that the player won't fall off when releasing the sneak key. The doors use a very minimalistic collision box, where the player is now centered on.