@paramat 2017-04-08
```
+				// Detect sneak ladder:
+				// Node two meters above sneak node must be solid
+				node = map->getNodeNoEx(m_sneak_node + v3s16(0,2,0),
+					&is_valid_position);
+				m_sneak_ladder_detected = is_valid_position &&
+					nodemgr->get(node).walkable;
```
Might be too simple and triggered by non-sneak ladders.
To make climb reasonable, the node 3 nodes above sneak node must not be walkable, otherwise there is no ledge to grab on to.