@paramat 2017-04-08
> Reduces climbing speed of sneak ladders !
1-row sneak ladders are slower than using regular ladders

The idea of new sneak code is to replicate old sneak ladder / jump behaviour, so climb speed must be similar to before and be unreasonably fast.