@SmallJoker 2017-04-08
@paramat I believe that the players would rather accept the sneak ladders if they weren't somewhat fast like now. It is not clear what the actual targets are. Is sneak now going to be fixed, changed or kept the same way? To decide this we need more player feedbacks about their idea of the sneak behaviour.

Damage evading, even on sneak ladders, is no longer possible. Some may disagree with this change too since it breaks the old, cheating behaviour.
If there are Windows users who would like to test it: [32-Bit binary,  d5aa9a0](https://www.dropbox.com/s/oco1df1q68xukkw/minetest_sneak_5533_170408_c28a843_3.7z) ([msvcr100.dll error fix](https://www.microsoft.com/de-ch/download/details.aspx?id=5555))