@paramat 2017-04-08
It's clear from the discussions that the players that like old sneak generally want sneak ladders to be as fast as they were before, because they like the mobility of the old code. They want the new code to have an option (now switched using 'sneak glitch') that replicates sneak ladder and sneak jump as closely as possible. The discussions have made this very clear and i think there has been more than enough player feedback over the last few weeks.
Are the slower sneak ladders in your code unavoidable?