@Fixer-007 2017-04-08
Damage evasion fixes? Very interesting, will try.