@Fixer-007 2017-04-08
Damage evasion like I described here https://github.com/minetest/minetest/issues/329#issuecomment-289164689 and https://github.com/minetest/minetest/issues/329#issuecomment-289311843 is still possible.