@SmallJoker 2017-04-11
1-row sneak ladder is now possible since #5527 - also trimmed this one to a minimum.
Did quite a few tests with it and it all works as expected. Am I missing something?