@paramat 2017-04-12
Will test.
I do feel the original high sneak ladder speed will generally be wanted, it's good to see this is possible if wanted.