@paramat 2017-04-13
Maybe leave docs until last to avoid conflicts?