@paramat 2017-04-13
Testing.

Sneak-jump:
Has quite a large and jolting teleport effect near the top of the jump similar to my first attempt at sneak-jump. If you release sneak and jump keys just before the teleport you end up hovering in this position unable to jump onto the ledge;

![screenshot_20170413_161300](https://cloud.githubusercontent.com/assets/3686677/25011183/6899f090-2064-11e7-9169-798d491b3f5d.png)

and then drop when using any move key.

1-column, 2-column and corner sneak ladders:
Is now a series of individual jumps instead of a smooth motion, so is slower than classic code. This repeated jumping doesn't feel good and will be irritating on long climbs.
Control on descent (holding and occasionally releasing sneak) is not as smooth and fine as current code, there is a 'bounce effect'.

It's now possible to sneak across the top of a stair node as if it were a full node, you can sneak above the 'missing part' of the stair node. This is not a big problem but is less realistic than current code.

> falling down if the node forward down is walkable but a nodebox.

^ Still present but not a problem.

> You can't walk across a default:snow or a 1/16 node while holding shift. To see this: drop a single node of default:snow on a flat surface, stand on the edge, hold shift and try to walk over it. You walk into an invisible wall and can't cross with shift held.

^ Still happens but invisible barrier is smaller, just like classic sneak code.

Jumping when at the limit of sneak works but jumping forwards has a much reduced forward speed so is not very useful.

@SmallJoker How do i enable faster climb on sneak ladders?