@paramat 2017-04-13
> Reduces climbing speed of sneak ladders slightly

Is an understatement. Players will not accept this slow series of jumps up a sneak ladder. We need to have a fast smooth climb as in current code.