@paramat 2017-04-13
With this PR i can't sneak on 1/32 slabs either above other nodes or not, but i can in current code.