@paramat 2017-04-13
> The idea behind this behaviour is basically:
    Higher attractivity of ladders and alternative transport
    Easier to tolerate for server admins (seriously, the current sneak ladder is just a fast elevator with damage-protection)
    Bonus: Closer to reality. Our super-Sam doesn't want to stay between two steps either

Yes the idea is fine in isolation, but we are trying to keep players happy who desperately want an option for behaviour as close as possible to classic sneak. Also i disagree with 'normalising' climbing made possible by holes in a wall 2m apart, it's too easy, if it's normalised it may be moved into default behaviour.