@paramat 2017-04-13
Testing with Line 266 changed for faster climb.
Simple sneak ladder is fast again but a series of abrupt teleports insted of a smooth motion, much faster than classic code now at 10n/s.
2-column and corner sneak ladders are the same but twice as fast, too fast now, about 20 nodes per second.
:-1: For current PR with either slow or fast sneak ladders, sorry.
However other advantages of your code may be good to have if you can add them to current sneak-jump and sneak-ladder behavioour.