@SmallJoker 2017-04-17
Measured the speed on a sneak ladder, 90 nodes upwards:
Custom fly speed: 4.59s (19.61 n/s) (`movement_speed_fast = 20`)
Old: 14.41s (6.25 n/s)
New: 12.76s (7.05 n/s), faster than original!
PR: 17.91s (5.02 n/s), still 2 n/s faster than climbing a regular ladder.
That's 20% slower than the original speed.
Test code: https://pastebin.com/MtRVGxYx