@paramat 2017-04-17
I think the speed depends on the type of ladder, with the single column ladder being slowest. Is this speed for the double-column ladders?