@paramat 2017-04-17
5 n/s seems ok but i still dislike the lack of smooth continuous movement and there are other issues.