@SmallJoker 2017-04-17
Measurements of 1-row sneak ladders (+/- compared with 2-row sneak ladders)
Old code: 12.71s (7.08 n/s), +0.8 n/s
Current: 13.32s (6.76 n/s), -0.3 n/s
This PR: 17.88s (5.03 n/s), +0 n/s

Well, at least it's consistent with 2-row ones :/